Lưu lại công việc bàn giao: bác sĩ/KTV, số lượng thập phân, đơn nghỉ đột xuất
Toàn bộ thay đổi này do một phiên làm việc khác thực hiện và để lại trên đĩa mà
chưa commit. Phiên đó đã dừng, nên gom lại thành một commit để không mất — mã
nguồn đang chạy thật trong production và database đã đổi theo nó.

Kho:
- Đơn xuất có thêm Bác sĩ và Kỹ thuật viên, bắt buộc nhập, hiện trong tin nhắn
  chờ duyệt và tin đã duyệt (v25)
- Số lượng nhận số thập phân một chữ số, ví dụ Filler 0.5 (v23, v24)
- Sửa duyệt đơn, điều chuyển, đồng bộ Sheet, danh mục sản phẩm và dịch vụ
- Web Admin: màn hình quản lý kho và cấu hình nhóm

Lịch khách:
- Tab "Hoàn Tất Lịch" cho nhóm report, bổ sung ảnh trong 48 giờ
- Nhắc lịch thiếu ảnh sau 30 phút, có dấu bền vững completion_reminded_at (v22)

Chấm công:
- Đơn nghỉ đột xuất và đi muộn có hiệu lực ngay, chụp lại lịch cũ để khôi phục
  chính xác khi quản lý từ chối (v21)
- Chặn đăng ký hoặc sửa lịch của ngày đã qua

Năm migration v21..v25 kèm script chạy. Đã đối chiếu: cả năm ĐỀU đã áp dụng lên
database. Không commit thì lịch sử mã nguồn và schema lệch nhau, khôi phục về bản
cũ sẽ hỏng.

Kèm theo, hai việc dọn dẹp của phiên hiện tại:
- Khôi phục 16 ảnh minh chứng bị xoá khỏi đĩa nhưng còn trong git. Cả 16 đều đang
  được lịch khách trỏ tới, để mất là 16 link 404. Số lịch có link ảnh hỏng giảm
  từ 63 xuống 47; 47 file còn lại chưa từng được commit nên không lấy lại được.
- Thêm npm run test:timekeep-module vào check:timekeep.

KHÔNG commit file rác tên "999" ở thư mục gốc (80 byte, UTF-16, nội dung là một
chuỗi ngày tháng) — nhiều khả năng do gõ nhầm lệnh chuyển hướng.

Kiểm chứng: 155/155 test xanh, hai điểm khởi động đều qua node --check.

+++ domains/warehouse/infrastructure/google-sheet/service-order-sheet-sync.js
@@ -88,8 +88,8 @@ export function createServiceOrderSheetSync({ pool, moment, getDocById }) {
             ),
             pool.query(
                 `SELECT p.id, p.barcode, p.product_name,
-                        COALESCE(MAX(i.quantity) FILTER (WHERE i.branch = 'US'), 0)::int AS stock_us,
-                        COALESCE(MAX(i.quantity) FILTER (WHERE i.branch = 'UK'), 0)::int AS stock_uk
+                        COALESCE(MAX(i.quantity) FILTER (WHERE i.branch = 'US'), 0) AS stock_us,
+                        COALESCE(MAX(i.quantity) FILTER (WHERE i.branch = 'UK'), 0) AS stock_uk
                  FROM tk_products p
                  JOIN tk_warehouse_order_items oi ON oi.product_id = p.id
                  JOIN tk_warehouse_order_services os ON os.id = oi.order_service_id
